+++ src/pages/api/all.json.ts
@@ -1,4 +1,4 @@
 import type { APIRoute } from "astro";
-import { dictionary } from "../dictionary";
+import { dictionary } from "../../dictionary";
 
 export const GET: APIRoute = () => new Response(JSON.stringify(dictionary.words));
