+++ src/pages/api/chapters.json.ts
@@ -1,4 +1,4 @@
 import type { APIRoute } from "astro";
-import chapters from "../data/chapters.json";
+import chapters from "../../data/chapters.json";
 
 export const GET: APIRoute = () => new Response(JSON.stringify(chapters));
